[FEAT] add comment delete

+++ functions/api/routes/classroomPost/classroomPostDELETE.js
@@ -3,7 +3,7 @@ const util = require("../../../lib/util");
 const statusCode = require("../../../constants/statusCode");
 const responseMessage = require("../../../constants/responseMessage");
 const db = require("../../../db/db");
-const { classroomPostDB } = require("../../../db");
+const { classroomPostDB, commentDB } = require("../../../db");
 
 module.exports = async (req, res) => {
   const { postId } = req.params;
@@ -20,15 +20,36 @@ module.exports = async (req, res) => {
 
     // 삭제하려는 유저와 게시글의 작성자가 같은지 확인
     let post = await classroomPostDB.getClassroomPostByPostId(client, postId);
+    console.log(post);
+    if (!post) {
+      return res
+        .status(statusCode.NOT_FOUND)
+        .send(util.fail(statusCode.NOT_FOUND, responseMessage.NO_POST));
+    }
     // 같지 않을 경우 403 FORBIDDEN Error
     if (post.writerId !== req.user.id) {
       return res
         .status(statusCode.FORBIDDEN)
-        .send(util.fail(statusCode.NOT_FOUND, responseMessage.FORBIDDEN_ACCESS));
+        .send(util.fail(statusCode.FORBIDDEN, responseMessage.FORBIDDEN_ACCESS));
     }
 
-    // 게시글 삭제하기
+    // 게시글 삭제
     let deletedPost = await classroomPostDB.deleteClassroomPostByPostId(client, postId);
+
+    if (!deletedPost) {
+      return res
+        .status(statusCode.NOT_FOUND)
+        .send(util.fail(statusCode.NOT_FOUND, responseMessage.NO_POST));
+    }
+
+    // 관련된 댓글 삭제
+    const deletedComment = await commentDB.deleteCommentByPostId(client, postId);
+
+    if (!deletedComment) {
+      return res
+        .status(statusCode.NOT_FOUND)
+        .send(util.fail(statusCode.NOT_FOUND, responseMessage.NO_COMMENT));
+    }
 
     res
       .status(statusCode.OK)
